# Conflicts:
#	include/vcpkg/commands.setinstalled.h
#	locales/messages.en.json
#	locales/messages.json
#	src/vcpkg/commands.porthistory.cpp
#	src/vcpkg/commands.setinstalled.cpp
#	src/vcpkg/install.cpp

+++ azure-pipelines/end-to-end-tests-dir/binarycaching.ps1
@@ -14,7 +14,7 @@ Run-Vcpkg -TestArgs ($commonArgs + @("remove", "rapidjson", "vcpkg-cmake", "vcpk
 Throw-IfFailed
 Require-FileNotExists "$installRoot/$Triplet/include"
 
-if(-Not $IsLinux) {
+if(-Not $IsLinux -and -Not $IsMacOS) {
     # Test simple nuget installation
     Run-Vcpkg -TestArgs ($commonArgs + @("install", "rapidjson", "vcpkg-cmake", "vcpkg-cmake-config", "--binarycaching", "--x-binarysource=clear;nuget,$NuGetRoot,readwrite"))
     Throw-IfFailed
@@ -47,7 +47,7 @@ Require-FileExists "$installRoot/$Triplet/include/rapidjson/rapidjson.h"
 Require-FileExists "$buildtreesRoot/rapidjson/src"
 Require-FileNotExists "$buildtreesRoot/detect_compiler"
 
-if(-Not $IsLinux) {
+if(-Not $IsLinux -and -Not $IsMacOS) {
     # Test restoring from nuget
     Remove-Item -Recurse -Force $installRoot
     Remove-Item -Recurse -Force $buildtreesRoot
